Enable floating point unit tests

+++ SingleSource/Regression/C/2003-05-21-UnionBitfields.c
@@ -6,7 +6,7 @@ int target_isinf(double x) {
   union {
     double d;
     struct {
-      unsigned mantissa2;
+      unsigned long mantissa2;
       unsigned long mantissa1 : 20;
       unsigned exponent  : 11;
       unsigned sign      :  1;
