+++ SingleSource/UnitTests/2005-07-17-INT-To-FP.c
@@ -1,7 +1,7 @@
 // Test the various SINT-TO-FP and UINT-TO-FP conversions.
 #include <stdio.h>
 
-int tests[] = {
+long tests[] = {
   0x80000000,
   -123456792,
   -10,
@@ -33,12 +33,12 @@ int main() {
            (double)(unsigned long long)((unsigned long long)i << 58ULL));  // i64
   }
   // edge case tests
-  for (i = 0; i < (sizeof(tests) / sizeof(int)); i++) {
+  for (i = 0; i < (sizeof(tests) / sizeof(long)); i++) {
     printf("%d %f, %f, %f, %f\n", i,
-                               (double)(unsigned)tests[i],
-                               (double)(  signed)tests[i],
+                               (double)(unsigned long)tests[i],
+                               (double)(  signed long)tests[i],
                                (float)           tests[i],
-                               (float) (  signed)tests[i]);
+                               (float) (  signed long)tests[i]);
   }
   return 0;
 }
